+++ dao-impl-db/src/main/java/com/yauhenikuntsevich/training/onlinestore/daodb/mapper/ProductMapper.java
@@ -14,7 +14,7 @@ public final class ProductMapper implements RowMapper<Product> {
 		Long product_id = rs.getLong("product_id");
 		Long categoryId = rs.getLong("category_id");
 		String name = rs.getString("name");
-		Integer price = rs.getInt("price");
+		Double price = rs.getDouble("price");
 		Integer quantity = rs.getInt("quantity");
 		String item = rs.getString("item");
